requirements.txt
fastapi
uvicorn
python-dotenv
requests
numpy
faiss-cpu
pdfplumber
google-generativeai
pydantic